Laravel 5.9 to remove helper string functions
In the next version of Laravel helper functions like `snake_case` do not work. If I recall there's an article someone on Laravel News talking about this change. I've been using the in a project that's using Laravel 5.9 and just ran into this today so I figured I'd send a PR.