Handle calling magic instantiation methods from within instance methods of the Enum
While it is not typical to use the magic instantiation dynamically, it may happen
incidentally when calling the instantiation in an instance method of itself.
Even when using the `static::KEY()` syntax, PHP still interprets this is a call to
an instance method when it happens inside of an instance method of the same class.